feat(google): Detect instance purchase option for Compute Addresses
Showing 3 cost components simultaneously for IP address is not correct,
if we can't determine which price to use, we should have a default case.

The usage param takes priority if set. If it's not set, we try to detect
the compute instance and determine its purchase option. If it's still
unclear we default to `unused` case.